Input Props isn't strictly a global, so move to its own file
