+++ src/globals.d.ts
@@ -27,14 +27,3 @@ declare module '*.svg' {
 	const content: any
 	export default content
 }
-
-/**
- * Mantine V5 removes this typing and returns an Any, so we need to override
- * that with what we actually recieve from the getInputProps function.
- */
-interface InputProps {
-	value: string
-	// eslint-disable-next-line @typescript-eslint/no-explicit-any
-	onChange: (event: any) => void
-	error: string
-}
